+++ pyawsmock/core.py
@@ -31,6 +31,10 @@ def client(service_name, region_name=None, **kwargs):
                 from pyawsmock.mocks.storage.s3.mock import MockS3
 
                 return MockS3(config.base_path)
+            elif service_name == "codeartifact":
+                from pyawsmock.mocks.developer_tools.codeartifact.mock import MockCodeArtifact
+
+                return MockCodeArtifact(config.base_path, region_name=region_name)
             else:
                 raise NotImplementedError(f"Local Mock not implemented for {service_name}")
         else: