+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Version History
 
+- `0.1.2`: Introcued CodeArtifact - `create_domain`, `list_domains`, `create_repository`, `delete_repository`,
+  `describe_repository`, `get_authorization_token`, `get_repository_endpoint`, `publish_package_version`,
+  `list_packages`, `describe_package`, `delete_package_versions`, `delete_package`, `list_package_versions`
 - `0.1.1`: Introduced S3 - `create_bucket`, `create_bucket_metadata_configuration`, `get_bucket_metadata_configuration`,
   `update_bucket_metadata_inventory_table_configuration`, `update_bucket_metadata_journal_table_configuration`,
   `delete_bucket_metadata_configuration`, `upload_file`, `download_file`, `get_object`